Restructured app so it's now a package not a model.
