+++ dopeticksApp/dopeticks/forms.py
@@ -1,6 +1,7 @@
 from flask_wtf import FlaskForm
 from wtforms import StringField, PasswordField, SubmitField, BooleanField
-from wtforms.validators import DataRequired, Length, Email, email_validator, EqualTo
+from wtforms.validators import DataRequired, Length, Email, email_validator, EqualTo, ValidationError
+from dopeticks.models import User
 
 
 
@@ -12,6 +13,14 @@ class RegistrationForm(FlaskForm):
 	userPassword = PasswordField('Password', validators=[DataRequired(), Length(min=8, max=20)])
 	confirmPassword = PasswordField('Confirm your password', validators=[DataRequired(), EqualTo('userPassword', message='Password must match')])
 	submit = SubmitField('Get Your Account!')
+
+	def validate_userEmail(self, userEmail):
+		user = User.query.filter_by(userEmail=userEmail.data).first()
+		# If the user query is none, nothing happens. Otherwise if the query returns data, throw validation error message
+		if user:
+			raise ValidationError("That email already exists. Please register with another email address")
+
+
 
 
 class LoginForm(FlaskForm):
