+++ dopeticksApp/dopeticks/templates/dashboard.html
@@ -1,0 +1,4 @@
+{% extends "layout.html" %}
+{% block content %}
+	<h1> Kia ora {{ current_user.userFirstName }}!</h1>
+{% endblock content %}
