+++ dopeticksApp/dopeticks/templates/home.html
