+++ dopeticksApp/dopeticks/templates/layout.html
@@ -21,7 +21,7 @@
 		<header class="site-header">
 		  <nav class="navbar navbar-expand-md navbar-dark bg-steel fixed-top">
 		    <div class="container">
-		      <a class="navbar-brand mr-4" href="/">Flask Blog</a>
+		      <a class="navbar-brand mr-4" href="/">Dopeticks</a>
 		      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarToggle" aria-controls="navbarToggle" aria-expanded="false" aria-label="Toggle navigation">
 		        <span class="navbar-toggler-icon"></span>
 		      </button>
@@ -32,8 +32,14 @@
 		        </div>
 		        <!-- Navbar Right Side -->
 		        <div class="navbar-nav">
-		          <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
-		          <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
+		        	<!-- # Using ginja2 conditional and using current_user.is_authenticated to check if user is logged in or not -->
+		        	{% if current_user.is_authenticated %}
+		        		<a class="nav-item nav-link" href="{{ url_for('dashboard') }}">{{ current_user.userFirstName }} {{ current_user.userLastName }}</a>
+		        		<a class="nav-item nav-link" href="{{ url_for('logout') }}">Logout</a>
+		        	{% else %}
+			          <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
+			          <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
+		          {% endif %}
 		        </div>
 		      </div>
 		    </div>
