+++ dopeticksApp/dopeticks/templates/login.html
@@ -48,7 +48,7 @@
 	</div>
 	<div class="border-top pt-3">
 		<small class="text-muted">
-			Need have an account? <a class="ml-2" href="{{ url_for('login') }}">Sign up here</a>
+			Want an account? <a class="ml-2" href="{{ url_for('login') }}">Sign up here</a>
 		</small>
 	</div>
 {% endblock content %}
