+++ dopeticksApp/requirements.txt
@@ -1,4 +1,5 @@
 awsebcli==3.20.5
+bcrypt==4.0.1
 botocore==1.29.81
 cement==2.8.2
 certifi==2022.12.7
@@ -10,6 +11,8 @@ dnspython==2.2.1
 email-validator==1.3.1
 filelock==3.8.2
 Flask==2.2.2
+Flask-Bcrypt==1.0.1
+Flask-Login==0.6.2
 Flask-PyMongo==2.3.0
 Flask-SQLAlchemy==3.0.3
 Flask-WTF==1.1.1
